+++ tools/interactive-evdev.c
@@ -146,7 +146,7 @@ keyboard_new(struct dirent *ent,
 
     const struct xkb_state_components_update components = {
         .size = sizeof(components),
-        .components = XKB_STATE_CONTROLS,
+        .components = XKB_STATE_CONTROLS_EFFECTIVE,
         .affect_controls = options->controls.boolean.affect_flags,
         .controls = options->controls.boolean.flags,
     };